+++ src/relation/transforms.rs
@@ -12,15 +12,74 @@ use crate::{
     },
     expr::{aggregate, Aggregate, Expr, Value},
     hierarchy::Hierarchy,
-    relation::Field,
-    DataType,
+    relation, DataType,
 };
 use itertools::Itertools;
+use sqlparser::test_utils::join;
 use std::collections::HashMap;
 use std::{
+    convert::Infallible,
+    error, fmt,
+    num::ParseFloatError,
     ops::{self, Deref},
     rc::Rc,
+    result,
 };
+
+#[derive(Debug, PartialEq)]
+pub enum Error {
+    InvalidRelation(String),
+    InvalidArguments(String),
+    Other(String),
+}
+
+impl Error {
+    pub fn invalid_relation(relation: impl fmt::Display) -> Error {
+        Error::InvalidRelation(format!("{} is invalid", relation))
+    }
+}
+
+impl fmt::Display for Error {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Error::InvalidRelation(desc) => writeln!(f, "InvalidRelation: {}", desc),
+            Error::InvalidArguments(desc) => writeln!(f, "InvalidArguments: {}", desc),
+            Error::Other(err) => writeln!(f, "{}", err),
+        }
+    }
+}
+
+impl error::Error for Error {}
+
+impl From<relation::Error> for Error {
+    fn from(err: relation::Error) -> Self {
+        Error::Other(err.to_string())
+    }
+}
+impl From<crate::expr::Error> for Error {
+    fn from(err: crate::expr::Error) -> Self {
+        Error::Other(err.to_string())
+    }
+}
+impl From<crate::io::Error> for Error {
+    fn from(err: crate::io::Error) -> Self {
+        Error::Other(err.to_string())
+    }
+}
+
+impl From<ParseFloatError> for Error {
+    fn from(err: ParseFloatError) -> Self {
+        Error::Other(err.to_string())
+    }
+}
+
+impl From<Infallible> for Error {
+    fn from(err: Infallible) -> Self {
+        Error::Other(err.to_string())
+    }
+}
+
+pub type Result<T> = result::Result<T, Error>;
 
 /* Reduce
  */
@@ -119,8 +178,12 @@ impl Reduce {
         self
     }
 
-    pub fn clip_aggregates(self, vectors: &str, clipping_values: Vec<(&str, f64)>) -> Relation {
-        let (map_names, vectors, base, coordinates): (
+    pub fn clip_aggregates(
+        self,
+        vectors: &str,
+        clipping_values: Vec<(&str, f64)>,
+    ) -> Result<Relation> {
+        let (map_names, out_vectors, base, coordinates): (
             Vec<(String, String)>,
             Option<String>,
             Vec<String>,
@@ -158,15 +221,29 @@ impl Reduce {
                 }
             });
 
-        assert_eq!(clipping_values.len(), coordinates.len());
+        let vectors = if let Some(v) = out_vectors {
+            Ok(v)
+        } else {
+            Err(Error::InvalidArguments(format!(
+                "{vectors} should be in the input `Relation`"
+            )))
+        };
+        let len_clipping_values = clipping_values.len();
+        let len_coordinates = coordinates.len();
+        if len_clipping_values != len_coordinates {
+            return Err(Error::InvalidArguments(format!(
+                "You must provide one clipping_value for each output field. \n \
+                Got {len_clipping_values} clipping values for {len_coordinates} output fields"
+            )));
+        }
         let clipped_relation = self.input.as_ref().clone().clipped_sum(
-            vectors.unwrap().as_str(),
+            vectors?.as_str(),
             base.iter().map(|s| s.as_str()).collect(),
             coordinates.iter().map(|s| s.as_str()).collect(),
             clipping_values,
         );
         let map_names: HashMap<String, String> = map_names.into_iter().collect();
-        clipped_relation.rename_fields(|n, _| map_names[n].to_string())
+        Ok(clipped_relation.rename_fields(|n, _| map_names[n].to_string()))
     }
 
     /// Rename fields
@@ -684,7 +761,7 @@ impl Relation {
         weighted_relation.sum_by(base, coordinates)
     }
 
-    pub fn clip_aggregates(self, vectors: &str, clipping_values: Vec<(&str, f64)>) -> Self {
+    pub fn clip_aggregates(self, vectors: &str, clipping_values: Vec<(&str, f64)>) -> Result<Self> {
         match self {
             Relation::Reduce(reduce) => reduce.clip_aggregates(vectors, clipping_values),
             _ => todo!(),
@@ -721,6 +798,27 @@ impl Relation {
             .filter(predicate)
             .input(self)
             .build()
+    }
+
+    /// Returns a filtered `Relation`
+    ///
+    /// # Arguments
+    /// - `columns`: `Vec<(column_name, minimal_value, maximal_value, possible_values)>`
+    ///
+    /// For example,
+    /// `filter_columns(vec![("my_col", Value::float(2.), Value::float(10.), vec![Value::integer(4), Value::integer(9)])])`
+    /// returns a filtered `Relation` whose `filter` is equivalent to `(my_col > 2.) and (my_col < 10) and (my_col in (4, 9)`
+    pub fn filter_columns(
+        self,
+        columns: Vec<(
+            &str,
+            Option<data_type::value::Value>,
+            Option<data_type::value::Value>,
+            Vec<data_type::value::Value>,
+        )>,
+    ) -> Relation {
+        let predicate = Expr::filter(columns);
+        self.filter(predicate)
     }
 
     /// Poisson sampling of a relation. It samples each line with probability 0 <= proba <= 1
@@ -776,6 +874,79 @@ impl Relation {
             .build();
         sampled_relation
     }
+
+    /// Returns a Relation whose fields have unique values
+    fn unique(self, columns: Vec<&str>) -> Relation {
+        let named_columns: Vec<(&str, Expr)> =
+            columns.into_iter().map(|c| (c, Expr::col(c))).collect();
+
+        Relation::reduce()
+            .group_by_iter(named_columns.iter().cloned().map(|(_, col)| col))
+            .with_iter(
+                named_columns
+                    .into_iter()
+                    .map(|(name, col)| (name, Expr::first(col))),
+            )
+            .input(self)
+            .build()
+    }
+
+    /// Returns a `Relation` whose output fields correspond to the `aggregates`
+    /// grouped by the expressions in `grouping_exprs`.
+    /// If `grouping_exprs` is not empty, we order by the grouping expressions.
+    fn build_ordered_reduce(
+        self,
+        grouping_exprs: Vec<Expr>,
+        aggregates: Vec<(&str, Expr)>,
+    ) -> Relation {
+        let red: Relation = Relation::reduce()
+            .with_iter(aggregates.clone())
+            .group_by_iter(grouping_exprs.clone())
+            .input(self)
+            .build();
+
+        if grouping_exprs.is_empty() {
+            red
+        } else {
+            Relation::map()
+                .with_iter(aggregates.into_iter().map(|(f, _)| (f, Expr::col(f))))
+                .order_by_iter(grouping_exprs.into_iter().map(|x| (x, true)).collect())
+                .input(red)
+                .build()
+        }
+    }
+
+    /// Build a relation whose output fields are to the aggregations in `aggregates`
+    /// applied on the UNIQUE values of the column `column` and grouped by the columns in `group_by`.
+    /// If `grouping_by` is not empty, we order by the grouping expressions.
+    pub fn distinct_aggregates(
+        self,
+        column: &str,
+        group_by: Vec<&str>,
+        aggregates: Vec<(&str, aggregate::Aggregate)>,
+    ) -> Relation {
+        let mut columns = vec![column];
+        columns.extend(group_by.iter());
+        let red = self.unique(columns);
+
+        // Build the second reduce
+        let mut aggregates_exprs: Vec<(&str, Expr)> = vec![];
+        let mut grouping_exprs: Vec<Expr> = vec![];
+        group_by.into_iter().for_each(|c| {
+            let col = Expr::col(c);
+            aggregates_exprs.push((c, Expr::first(col.clone())));
+            grouping_exprs.push(col);
+        });
+        aggregates.into_iter().for_each(|(c, agg)| {
+            aggregates_exprs.push((
+                c,
+                Expr::Aggregate(Aggregate::new(agg, Rc::new(Expr::col(column)))),
+            ))
+        });
+
+        // Add order by
+        red.build_ordered_reduce(grouping_exprs, aggregates_exprs)
+    }
 }
 
 impl With<(&str, Expr)> for Relation {
@@ -792,10 +963,12 @@ mod tests {
         data_type::{value::List, DataTyped},
         display::Dot,
         io::{postgresql, Database},
+        relation::schema::Schema,
         sql::parse,
     };
     use colored::Colorize;
     use itertools::Itertools;
+    use sqlparser::keywords::RIGHT;
 
     #[test]
     fn test_with_computed_field() {
@@ -894,8 +1067,7 @@ mod tests {
         for row in results {
             let mut str_row = vec![];
             for i in 0..size {
-                let float_i: Result<f64, _> = row[i].to_string().parse();
-                str_row.push(match float_i {
+                str_row.push(match row[i].to_string().parse::<f64>() {
                     Ok(f) => ((f * 1000.).round() / 1000.).to_string(),
                     Err(_) => row[i].to_string(),
                 })
@@ -1237,7 +1409,9 @@ mod tests {
 
         let schema = my_relation.inputs()[0].schema().clone();
         let price = schema.field_from_index(0).unwrap().name();
-        let clipped_relation = my_relation.clip_aggregates("order_id", vec![(price, 45.)]);
+        let clipped_relation = my_relation
+            .clip_aggregates("order_id", vec![(price, 45.)])
+            .unwrap();
         let name_fields: Vec<&str> = clipped_relation.schema().iter().map(|f| f.name()).collect();
         assert_eq!(name_fields, vec!["item", "sum_price"]);
         clipped_relation.display_dot();
@@ -1265,7 +1439,9 @@ mod tests {
 
         let schema = my_relation.inputs()[0].schema().clone();
         let price = schema.field_from_index(0).unwrap().name();
-        let clipped_relation = my_relation.clip_aggregates("order_id", vec![(price, 45.)]);
+        let clipped_relation = my_relation
+            .clip_aggregates("order_id", vec![(price, 45.)])
+            .unwrap();
         let name_fields: Vec<&str> = clipped_relation.schema().iter().map(|f| f.name()).collect();
         assert_eq!(name_fields, vec!["sum_price"]);
         clipped_relation.display_dot();
@@ -1306,8 +1482,9 @@ mod tests {
         let schema = relation.inputs()[0].schema().clone();
         let price = schema.field_from_index(2).unwrap().name();
         let std_price = schema.field_from_index(3).unwrap().name();
-        let clipped_relation =
-            relation.clip_aggregates("user_id", vec![(price, 45.), (std_price, 50.)]);
+        let clipped_relation = relation
+            .clip_aggregates("user_id", vec![(price, 45.), (std_price, 50.)])
+            .unwrap();
         clipped_relation.display_dot();
         let name_fields: Vec<&str> = clipped_relation.schema().iter().map(|f| f.name()).collect();
         assert_eq!(name_fields, vec!["item", "sum1", "sum2"]);
@@ -1689,5 +1866,100 @@ mod tests {
             (&query_sampled_relation[..]).replace(' ', "")
         );
         print!("{}\n", query_sampled_relation)
+    }
+
+    #[test]
+    fn test_unique() {
+        let table: Relation = Relation::table()
+            .name("table")
+            .schema(
+                Schema::builder()
+                    .with(("a", DataType::integer_range(1..=10)))
+                    .with(("b", DataType::integer_values([1, 2, 5, 6, 7, 8])))
+                    .with(("c", DataType::integer_range(5..=20)))
+                    .build(),
+            )
+            .build();
+
+        // Without group by
+        let unique_rel = table.unique(vec!["a", "b"]);
+        println!("{}", unique_rel);
+        _ = unique_rel.display_dot();
+    }
+
+    #[test]
+    fn test_build_ordered_reduce() {
+        let table: Relation = Relation::table()
+            .name("table")
+            .schema(
+                Schema::builder()
+                    .with(("a", DataType::integer_range(1..=10)))
+                    .with(("b", DataType::integer_values([1, 2, 5, 6, 7, 8])))
+                    .with(("c", DataType::integer_range(5..=20)))
+                    .build(),
+            )
+            .build();
+
+        // Without group by
+        let grouping_exprs = vec![];
+        let aggregates = vec![
+            ("sum_a", Expr::sum(Expr::col("a"))),
+            ("count_b", Expr::count(Expr::col("a"))),
+        ];
+        let rel = table
+            .clone()
+            .build_ordered_reduce(grouping_exprs, aggregates);
+        println!("{}", rel);
+        _ = rel.display_dot();
+
+        // With group by
+        let grouping_exprs = vec![Expr::col("c")];
+        let aggregates = vec![
+            ("sum_a", Expr::sum(Expr::col("a"))),
+            ("count_b", Expr::count(Expr::col("a"))),
+        ];
+        let rel = table.build_ordered_reduce(grouping_exprs, aggregates);
+        println!("{}", rel);
+        _ = rel.display_dot();
+    }
+
+    #[test]
+    fn test_distinct_aggregates() {
+        let table: Relation = Relation::table()
+            .name("table")
+            .schema(
+                Schema::builder()
+                    .with(("a", DataType::integer_range(1..=10)))
+                    .with(("b", DataType::integer_values([1, 2, 5, 6, 7, 8])))
+                    .with(("c", DataType::integer_range(5..=20)))
+                    .build(),
+            )
+            .build();
+
+        // Without group by
+        let column = "a";
+        let group_by = vec![];
+        let aggregates = vec![
+            ("sum_distinct_a", aggregate::Aggregate::Sum),
+            ("count_distinct_a", aggregate::Aggregate::Count),
+        ];
+        let distinct_rel = table
+            .clone()
+            .distinct_aggregates(column, group_by, aggregates);
+        println!("{}", distinct_rel);
+        _ = distinct_rel.display_dot();
+
+        // With group by
+        let column = "a";
+        let group_by = vec!["b", "c"];
+        let aggregates = vec![
+            ("sum_distinct_a", aggregate::Aggregate::Sum),
+            ("count_distinct_a", aggregate::Aggregate::Count),
+        ];
+        let distinct_rel = table
+            .clone()
+            .distinct_aggregates(column, group_by, aggregates);
+        println!("{}", distinct_rel);
+        _ = distinct_rel.display_dot();
     }
 }